added paint

+++ cpp-sources/construct/matrix.hh
@@ -19,7 +19,7 @@ public:
   number max();
 
   matrix<number> *transpose();
-  matrix<uint8_t> *non_zero_transposed();
+  matrix<int> *non_zero_transposed();
 
   matrix<number> *operator*(float n);
   matrix<number> *operator/(float n);
@@ -39,6 +39,6 @@ template <typename num1, typename num2>
 matrix<num1> *mat_add_element_wise(matrix<num1> *mat1, matrix<num2> *mat2);
 
 template <typename number>
-void quickSort(matrix<uint8_t> *indices, matrix<number> *values, int start, int end);
+void quickSort(matrix<int> *indices, matrix<number> *values, int start, int end);
 
 #endif /* MATRIX_HH */
