+++ cpp-sources/tools/harris.hh
@@ -6,5 +6,5 @@
 
 matrix<float> *compute_harris_response(matrix<uint8_t> *img);
 
-matrix<uint8_t> *detect_harris_points(matrix<uint8_t> *image_gray, int max_keypoints=30, 
+matrix<int> *detect_harris_points(matrix<uint8_t> *image_gray, int max_keypoints=30, 
 int min_distance=25, float threshold=0.1);
